build: Fix qemugdb target
Even deprecated, this target should still work, the issue was
introduced by commit 7be222e8de5fb7ca727d66b8de470d81cd46f57b [formerly 2bc9d699817e611dbe144f332fb7067f540b8e19].

The problem is that for both of those targets the board's specific
Makefiles should be include'd, it was done only for the non-deprecated
target ('debugserver').

As a note: that is the recommended way of doing a logical OR operation
in Makefiles.

Change-Id: I3ae8f5201c47e65b33a62cea45e25dc2226de489
Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>

Former-commit-id: 370571b563a5c9940aa65288801cd5e35d7ecab3